+++ src/express/middleware/error.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import HttpStatus from 'http-status-codes';
-import { Logger } from '../../utils/logging';
-import { HttpError, InternalServerError } from '../errors';
-import { createErrorResponse } from './../response';
+import { Logger } from '@utils/logging';
+import { HttpError, InternalServerError } from '@express/errors';
+import { createErrorResponse } from '@express/response';
 
 export const handleError = (logger: Logger) => (error: Error, _: Request, res: Response, next: NextFunction) => {
   const httpError: HttpError = error instanceof HttpError ? error : new InternalServerError(error.message);
