+++ src/workflow/Workflow.ts
@@ -1,7 +1,7 @@
-import { Application } from '../Application';
-import { Trigger } from './Trigger';
-import { Logger } from './../utils/logging';
-import { createExecutionTimer } from '../utils/time';
+import { Application } from '@application/Application';
+import { Trigger } from '@workflow/Trigger';
+import { Logger } from '@utils/logging';
+import { createExecutionTimer } from '@utils/time';
 
 export interface WorkflowMetadata {
   name: string
@@ -36,7 +36,7 @@ export abstract class Workflow<T> {
       const timer = createExecutionTimer();
       await this.run(payload);
       const executionDuration = timer();
-      
+
       this.logger.info(`Workflow has finished running. (Execution took ${executionDuration.formattedDuration})`);
       this.application.emit('workflowFinish', this);
     } catch (error) {
