+++ src/workflow/triggers/IntervalTrigger.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-import { Trigger, Clearable } from '../Trigger';
+import { Trigger, Clearable } from '@workflow/Trigger';
 
 export class IntervalTrigger extends Trigger<void> implements Clearable {
   private handler: NodeJS.Timer | null;
