+++ src/workflow/triggers/TimeoutTrigger.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-undef */
-import { Trigger, Clearable } from '../Trigger';
+import { Trigger, Clearable } from '@workflow/Trigger';
 
 export class TimeoutTrigger extends Trigger<void> implements Clearable {
   private handler: NodeJS.Timeout | null;
