+++ workflows/media/TautulliNotificationsWorkflow.ts
@@ -1,8 +1,8 @@
 /* eslint-disable max-statements */
-import { Workflow } from '../../Workflow';
-import { Application } from '../../../Application';
-import * as DiscordWebhook from '../../../clients/discordWebhook';
-import { config } from '../../../config';
+import { Workflow } from '@workflow/Workflow';
+import { Application } from '@application/Application';
+import * as DiscordWebhook from '@clients/discordWebhook';
+import { config } from '@config/config';
 
 const EMBED_COLOR = 15048717;
 
@@ -196,7 +196,7 @@ export class TautulliNotificationsWorkflow extends Workflow<TriggerPayload> {
         }]
       };
     }
-    
+
     return null;
   }
 }
