+++ workflows/twitter/AutoRetweetWorkflow.ts
@@ -1,9 +1,9 @@
 import { TweetV2SingleStreamResult } from 'twitter-api-v2';
-import { Workflow } from '../../Workflow';
-import { Application } from './../../../Application';
-import * as Twitter from '../../../clients/twitter';
-import { levelDatabaseService } from '../../../services/LevelDatabaseService';
-import { config } from '../../../config';
+import { Workflow } from '@workflow/Workflow';
+import { Application } from '@application/Application';
+import * as Twitter from '@clients/twitter';
+import { levelDatabaseService } from '@services/LevelDatabaseService';
+import { config } from '@config/config';
 
 export class AutoRetweetWorkflow extends Workflow<TweetV2SingleStreamResult> {
   private twitterClient: Twitter.ClientV1;
